Struct muda::accelerator::Accelerator

pub struct Accelerator { /* private fields */ }

A keyboard shortcut that consists of an optional combination of modifier keys (provided by Modifiers and one key (Code).

Implementations

impl Accelerator

pub fn new(mods: Option<Modifiers>, key: Code) -> Self

Creates a new accelerator to define keyboard shortcuts throughout your application. Only Modifiers::ALT, Modifiers::SHIFT, Modifiers::CONTROL, and Modifiers::SUPER

pub fn id(&self) -> u32

Returns the id associated with this accelerator which is a hash of the string representation of modifiers and key within this accelerator.

pub fn matches( &self, modifiers: impl Borrow<Modifiers>, key: impl Borrow<Code>, ) -> bool

Returns true if this Code and Modifiers matches this Accelerator.

impl Accelerator

pub fn key_equivalent(self) -> Result<String, AcceleratorParseError>

Return the string value of this hotkey, without modifiers.

Returns the empty string if no key equivalent is known.

pub fn key_modifier_mask(self) -> NSEventModifierFlags

Return the modifiers of this hotkey, as an NSEventModifierFlags bitflag.