muda

Enum NativeIcon

pub enum NativeIcon {

    Add,
    Advanced,
    Bluetooth,
    Bookmarks,
    Caution,
    ColorPanel,
    ColumnView,
    Computer,
    EnterFullScreen,
    Everyone,
    ExitFullScreen,
    FlowView,
    Folder,
    FolderBurnable,
    FolderSmart,
    FollowLinkFreestanding,
    FontPanel,
    GoLeft,
    GoRight,
    Home,
    IChatTheater,
    IconView,
    Info,
    InvalidDataFreestanding,
    LeftFacingTriangle,
    ListView,
    LockLocked,
    LockUnlocked,
    MenuMixedState,
    MenuOnState,
    MobileMe,
    MultipleDocuments,
    Network,
    Path,
    PreferencesGeneral,
    QuickLook,
    RefreshFreestanding,
    Refresh,
    Remove,
    RevealFreestanding,
    RightFacingTriangle,
    Share,
    Slideshow,
    SmartBadge,
    StatusAvailable,
    StatusNone,
    StatusPartiallyAvailable,
    StatusUnavailable,
    StopProgressFreestanding,
    StopProgress,
    TrashEmpty,
    TrashFull,
    User,
    UserAccounts,
    UserGroup,
    UserGuest,

}

A native Icon to be used for the menu item

Platform-specific:

Windows / Linux: Unsupported.

Variants

Add

An add item template image.

Advanced

Advanced preferences toolbar icon for the preferences window.

Bluetooth

A Bluetooth template image.

Bookmarks

Bookmarks image suitable for a template.

Caution

A caution image.

ColorPanel

A color panel toolbar icon.

ColumnView

A column view mode template image.

Computer

A computer icon.

EnterFullScreen

An enter full-screen mode template image.

Everyone

Permissions for all users.

ExitFullScreen

An exit full-screen mode template image.

FlowView

A cover flow view mode template image.

Folder

A folder image.

FolderBurnable

A burnable folder icon.

FolderSmart

A smart folder icon.

FollowLinkFreestanding

A link template image.

FontPanel

A font panel toolbar icon.

GoLeft

A go back template image.

GoRight

A go forward template image.

Home

Home image suitable for a template.

IChatTheater

An iChat Theater template image.

IconView

An icon view mode template image.

Info

An information toolbar icon.

InvalidDataFreestanding

A template image used to denote invalid data.

LeftFacingTriangle

A generic left-facing triangle template image.

ListView

A list view mode template image.

LockLocked

A locked padlock template image.

LockUnlocked

An unlocked padlock template image.

MenuMixedState

A horizontal dash, for use in menus.

MenuOnState

A check mark template image, for use in menus.

MobileMe

A MobileMe icon.

MultipleDocuments

A drag image for multiple items.

Network

A network icon.

Path

A path button template image.

PreferencesGeneral

General preferences toolbar icon for the preferences window.

QuickLook

A Quick Look template image.

RefreshFreestanding

A refresh template image.

Refresh

A refresh template image.

Remove

A remove item template image.

RevealFreestanding

A reveal contents template image.

RightFacingTriangle

A generic right-facing triangle template image.

Share

A share view template image.

Slideshow

A slideshow template image.

SmartBadge

A badge for a smart item.

StatusAvailable

Small green indicator, similar to iChatโs available image.

StatusNone

Small clear indicator.

StatusPartiallyAvailable

Small yellow indicator, similar to iChatโs idle image.

StatusUnavailable

Small red indicator, similar to iChatโs unavailable image.

StopProgressFreestanding

A stop progress template image.

StopProgress

A stop progress button template image.

TrashEmpty

An image of the empty trash can.

TrashFull

An image of the full trash can.

User

Permissions for a single user.

UserAccounts

User account toolbar icon for the preferences window.

UserGroup

Permissions for a group of users.

UserGuest

Permissions for guests.