Struct muda::CheckMenuItem

pub struct CheckMenuItem { /* private fields */ }

A check menu item inside a Menu or Submenu and usually contains a text and a check mark or a similar toggle that corresponds to a checked and unchecked states.

Implementations

impl CheckMenuItem

pub fn new<S: AsRef<str>>( text: S, enabled: bool, checked: bool, accelerator: Option<Accelerator>, ) -> Self

Create a new check menu item.

text could optionally contain an & before a character to assign this character as the mnemonic for this check menu item. To display a & without assigning a mnemenonic, use &&.

pub fn with_id<I: Into<MenuId>, S: AsRef<str>>( id: I, text: S, enabled: bool, checked: bool, accelerator: Option<Accelerator>, ) -> Self

Create a new check menu item with the specified id.

text could optionally contain an & before a character to assign this character as the mnemonic for this check menu item. To display a & without assigning a mnemenonic, use &&.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this submenu.

pub fn text(&self) -> String

Get the text for this check menu item.

pub fn set_text<S: AsRef<str>>(&self, text: S)

Set the text for this check menu item. text could optionally contain an & before a character to assign this character as the mnemonic for this check menu item. To display a & without assigning a mnemenonic, use &&.

pub fn is_enabled(&self) -> bool

Get whether this check menu item is enabled or not.

pub fn set_enabled(&self, enabled: bool)

Enable or disable this check menu item.

pub fn set_accelerator(&self, accelerator: Option<Accelerator>) -> Result<()>

Set this check menu item accelerator.

pub fn is_checked(&self) -> bool

Get whether this check menu item is checked or not.

pub fn set_checked(&self, checked: bool)

Check or Uncheck this check menu item.

pub fn into_id(self) -> MenuId

Convert this menu item into its menu ID.