Struct muda::Menu

pub struct Menu { /* private fields */ }

A root menu that can be added to a Window on Windows and Linux and used as the app global menu on macOS.

Implementations

impl Menu

pub fn new() -> Self

Creates a new menu.

pub fn with_id<I: Into<MenuId>>(id: I) -> Self

Creates a new menu with the specified id.

pub fn with_items(items: &[&dyn IsMenuItem]) -> Result<Self>

Creates a new menu with given items. It calls Menu::new and Menu::append_items internally.

pub fn with_id_and_items<I: Into<MenuId>>( id: I, items: &[&dyn IsMenuItem], ) -> Result<Self>

Creates a new menu with the specified id and given items. It calls Menu::new and Menu::append_items internally.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this menu.

pub fn append(&self, item: &dyn IsMenuItem) -> Result<()>

Add a menu item to the end of this menu.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn append_items(&self, items: &[&dyn IsMenuItem]) -> Result<()>

Add menu items to the end of this menu. It calls Menu::append in a loop internally.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn prepend(&self, item: &dyn IsMenuItem) -> Result<()>

Add a menu item to the beginning of this menu.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn prepend_items(&self, items: &[&dyn IsMenuItem]) -> Result<()>

Add menu items to the beginning of this menu. It calls Menu::insert_items with position of 0 internally.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn insert(&self, item: &dyn IsMenuItem, position: usize) -> Result<()>

Insert a menu item at the specified postion in the menu.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn insert_items( &self, items: &[&dyn IsMenuItem], position: usize, ) -> Result<()>

Insert menu items at the specified postion in the menu.

Platform-spcific:

macOS: Only Submenu can be added to the menu

pub fn remove(&self, item: &dyn IsMenuItem) -> Result<()>

Remove a menu item from this menu.

pub fn remove_at(&self, position: usize) -> Option<MenuItemKind>

Remove the menu item at the specified position from this menu and returns it.

pub fn items(&self) -> Vec<MenuItemKind>

Returns a list of menu items that has been added to this menu.

pub fn init_for_nsapp(&self)

Adds this menu to an NSApp.

pub fn remove_for_nsapp(&self)

Removes this menu from an NSApp.