Struct muda::PredefinedMenuItem

pub struct PredefinedMenuItem { /* private fields */ }

A predefined (native) menu item which has a predfined behavior by the OS or by this crate.

Implementations

impl PredefinedMenuItem

pub fn separator() -> PredefinedMenuItem

Separator menu item

pub fn copy(text: Option<&str>) -> PredefinedMenuItem

Copy menu item

pub fn cut(text: Option<&str>) -> PredefinedMenuItem

Cut menu item

pub fn paste(text: Option<&str>) -> PredefinedMenuItem

Paste menu item

pub fn select_all(text: Option<&str>) -> PredefinedMenuItem

SelectAll menu item

pub fn undo(text: Option<&str>) -> PredefinedMenuItem

Undo menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn redo(text: Option<&str>) -> PredefinedMenuItem

Redo menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn minimize(text: Option<&str>) -> PredefinedMenuItem

Minimize window menu item

Platform-specific:

Linux: Unsupported.

pub fn maximize(text: Option<&str>) -> PredefinedMenuItem

Maximize window menu item

Platform-specific:

Linux: Unsupported.

pub fn fullscreen(text: Option<&str>) -> PredefinedMenuItem

Fullscreen menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn hide(text: Option<&str>) -> PredefinedMenuItem

Hide window menu item

Platform-specific:

Linux: Unsupported.

pub fn hide_others(text: Option<&str>) -> PredefinedMenuItem

Hide other windows menu item

Platform-specific:

Linux: Unsupported.

pub fn show_all(text: Option<&str>) -> PredefinedMenuItem

Show all app windows menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn close_window(text: Option<&str>) -> PredefinedMenuItem

Close window menu item

Platform-specific:

Linux: Unsupported.

pub fn quit(text: Option<&str>) -> PredefinedMenuItem

Quit app menu item

Platform-specific:

Linux: Unsupported.

pub fn about( text: Option<&str>, metadata: Option<AboutMetadata>, ) -> PredefinedMenuItem

About app menu item

pub fn services(text: Option<&str>) -> PredefinedMenuItem

Services menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn bring_all_to_front(text: Option<&str>) -> PredefinedMenuItem

โBring all to frontโ menu item

Platform-specific:

Windows / Linux: Unsupported.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this predefined menu item.

pub fn text(&self) -> String

Get the text for this predefined menu item.

pub fn set_text<S: AsRef<str>>(&self, text: S)

Set the text for this predefined menu item.

pub fn into_id(self) -> MenuId

Convert this menu item into its menu ID.