Struct muda::Submenu

pub struct Submenu { /* private fields */ }

A menu that can be added to a Menu or another Submenu.

Implementations

impl Submenu

pub fn new<S: AsRef<str>>(text: S, enabled: bool) -> Self

Create a new submenu.

text could optionally contain an & before a character to assign this character as the mnemonic for this submenu. To display a & without assigning a mnemenonic, use &&.

pub fn with_id<I: Into<MenuId>, S: AsRef<str>>( id: I, text: S, enabled: bool, ) -> Self

Create a new submenu with the specified id.

text could optionally contain an & before a character to assign this character as the mnemonic for this submenu. To display a & without assigning a mnemenonic, use &&.

pub fn with_items<S: AsRef<str>>( text: S, enabled: bool, items: &[&dyn IsMenuItem], ) -> Result<Self>

Creates a new submenu with given items. It calls Submenu::new and Submenu::append_items internally.

pub fn with_id_and_items<I: Into<MenuId>, S: AsRef<str>>( id: I, text: S, enabled: bool, items: &[&dyn IsMenuItem], ) -> Result<Self>

Creates a new submenu with the specified id and given items. It calls Submenu::new and Submenu::append_items internally.

pub fn id(&self) -> &MenuId

Returns a unique identifier associated with this submenu.

pub fn append(&self, item: &dyn IsMenuItem) -> Result<()>

Add a menu item to the end of this menu.

pub fn append_items(&self, items: &[&dyn IsMenuItem]) -> Result<()>

Add menu items to the end of this submenu. It calls Submenu::append in a loop.

pub fn prepend(&self, item: &dyn IsMenuItem) -> Result<()>

Add a menu item to the beginning of this submenu.

pub fn prepend_items(&self, items: &[&dyn IsMenuItem]) -> Result<()>

Add menu items to the beginning of this submenu. It calls Menu::prepend on the first element and passes the rest to Menu::insert_items with position of 1.

pub fn insert(&self, item: &dyn IsMenuItem, position: usize) -> Result<()>

Insert a menu item at the specified postion in the submenu.

pub fn insert_items( &self, items: &[&dyn IsMenuItem], position: usize, ) -> Result<()>

Insert menu items at the specified postion in the submenu.

pub fn remove(&self, item: &dyn IsMenuItem) -> Result<()>

Remove a menu item from this submenu.

pub fn remove_at(&self, position: usize) -> Option<MenuItemKind>

Remove the menu item at the specified position from this submenu and returns it.

pub fn items(&self) -> Vec<MenuItemKind>

Returns a list of menu items that has been added to this submenu.

pub fn text(&self) -> String

Get the text for this submenu.

pub fn set_text<S: AsRef<str>>(&self, text: S)

Set the text for this submenu. text could optionally contain an & before a character to assign this character as the mnemonic for this submenu. To display a & without assigning a mnemenonic, use &&.

pub fn is_enabled(&self) -> bool

Get whether this submenu is enabled or not.

pub fn set_enabled(&self, enabled: bool)

Enable or disable this submenu.

pub fn set_as_windows_menu_for_nsapp(&self)

Set this submenu as the Window menu for the application on macOS.

This will cause macOS to automatically add window-switching items and certain other items to the menu.

pub fn set_as_help_menu_for_nsapp(&self)

Set this submenu as the Help menu for the application on macOS.

This will cause macOS to automatically add a search box to the menu.

If no menu is set as the Help menu, macOS will automatically use any menu which has a title matching the localized word โHelpโ.

pub fn into_id(self) -> MenuId

Convert this submenu into its menu ID.